extract use case
